# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/knowledge/cards:
        get:
            tags:
                - KnowledgeCardService
            operationId: KnowledgeCardService_ListKnowledgeCards
            parameters:
                - name: includeArchived
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListKnowledgeCardsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - KnowledgeCardService
            operationId: KnowledgeCardService_CreateKnowledgeCard
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateKnowledgeCardRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/KnowledgeCard'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        CardSourceSelection:
            type: object
            properties:
                document:
                    type: string
                contentHash:
                    type: string
                startOffset:
                    type: integer
                    format: int32
                endOffset:
                    type: integer
                    format: int32
...
        CreateKnowledgeCardRequest:
            required:
                - card
            type: object
            properties:
                card:
                    $ref: '#/components/schemas/KnowledgeCard'
                source:
                    $ref: '#/components/schemas/CardSourceSelection'
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        KnowledgeCard:
            required:
                - title
                - body
            type: object
            properties:
                name:
                    type: string
                title:
                    type: string
                body:
                    type: string
                cardType:
                    type: string
                topics:
                    type: array
                    items:
                        type: string
                archived:
                    type: boolean
                sourceDocument:
                    readOnly: true
                    type: string
                sourceTitle:
                    readOnly: true
                    type: string
                sourceQuote:
                    readOnly: true
                    type: string
                sourceLocator:
                    readOnly: true
                    type: string
                sourceContentHash:
                    readOnly: true
                    type: string
                createTime:
                    readOnly: true
                    type: string
                    format: date-time
                updateTime:
                    readOnly: true
                    type: string
                    format: date-time
                sourceAvailable:
                    readOnly: true
                    type: boolean
...
        ListKnowledgeCardsResponse:
            type: object
            properties:
                cards:
                    type: array
                    items:
                        $ref: '#/components/schemas/KnowledgeCard'
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...